+++ src/vimina.hpp
@@ -1,8 +1,28 @@
-#pragma once
+﻿#pragma once
 
 namespace vimina {
     static const std::vector<std::string> modeLabels = {
         "Clock multiplier/divider",
         "Clock swing"
     };
+
+    static const std::vector<std::string> factorLabels = {
+        "x1",
+        "x2",
+        "x3",
+        "x4",
+        "x5",
+        "x6",
+        "x7",
+        "x8",
+        "÷2",
+        "÷3",
+        "÷4",
+        "÷5",
+        "÷6",
+        "÷7",
+        "÷8"
+    };
+
+    static const std::string clockPrefix = "Clock: ";
 }